+++ lib/config/bigquery.go
@@ -2,15 +2,6 @@ package config
 
 import "fmt"
 
-type BigQuery struct {
-	// PathToCredentials is _optional_ if you have GOOGLE_APPLICATION_CREDENTIALS set as an env var
-	// Links to credentials: https://cloud.google.com/docs/authentication/application-default-credentials#GAC
-	PathToCredentials string `yaml:"pathToCredentials"`
-	DefaultDataset    string `yaml:"defaultDataset"`
-	ProjectID         string `yaml:"projectID"`
-	Location          string `yaml:"location"`
-}
-
 // DSN - returns the notation for BigQuery following this format: bigquery://projectID/[location/]datasetID?queryString
 // If location is passed in, we'll specify it. Else, it'll default to empty and our library will set it to US.
 func (b *BigQuery) DSN() string {
