+++ lib/config/destination_types.go
@@ -1,0 +1,53 @@
+package config
+
+import "github.com/artie-labs/transfer/lib/config/constants"
+
+type S3Settings struct {
+	FolderName         string                   `yaml:"folderName"`
+	Bucket             string                   `yaml:"bucket"`
+	AwsAccessKeyID     string                   `yaml:"awsAccessKeyID"`
+	AwsSecretAccessKey string                   `yaml:"awsSecretAccessKey"`
+	OutputFormat       constants.S3OutputFormat `yaml:"outputFormat"`
+}
+
+type MSSQL struct {
+	Host     string `yaml:"host"`
+	Port     int    `yaml:"port"`
+	Username string `yaml:"username"`
+	Password string `yaml:"password"`
+	Database string `yaml:"database"`
+}
+
+type BigQuery struct {
+	// PathToCredentials is _optional_ if you have GOOGLE_APPLICATION_CREDENTIALS set as an env var
+	// Links to credentials: https://cloud.google.com/docs/authentication/application-default-credentials#GAC
+	PathToCredentials string `yaml:"pathToCredentials"`
+	DefaultDataset    string `yaml:"defaultDataset"`
+	ProjectID         string `yaml:"projectID"`
+	Location          string `yaml:"location"`
+}
+
+type Redshift struct {
+	Host             string `yaml:"host"`
+	Port             int    `yaml:"port"`
+	Database         string `yaml:"database"`
+	Username         string `yaml:"username"`
+	Password         string `yaml:"password"`
+	Bucket           string `yaml:"bucket"`
+	OptionalS3Prefix string `yaml:"optionalS3Prefix"`
+	// https://docs.aws.amazon.com/redshift/latest/dg/copy-parameters-authorization.html
+	CredentialsClause string `yaml:"credentialsClause"`
+}
+
+type Snowflake struct {
+	AccountID string `yaml:"account"`
+	Username  string `yaml:"username"`
+	// If pathToPrivateKey is specified, the password field will be ignored
+	PathToPrivateKey string `yaml:"pathToPrivateKey,omitempty"`
+	Password         string `yaml:"password,omitempty"`
+
+	Warehouse   string `yaml:"warehouse"`
+	Region      string `yaml:"region"`
+	Host        string `yaml:"host"`
+	Application string `yaml:"application"`
+}
