+++ lib/config/mssql.go
@@ -8,14 +8,6 @@ import (
 	"github.com/artie-labs/transfer/lib/stringutil"
 )
 
-type MSSQL struct {
-	Host     string `yaml:"host"`
-	Port     int    `yaml:"port"`
-	Username string `yaml:"username"`
-	Password string `yaml:"password"`
-	Database string `yaml:"database"`
-}
-
 func (m MSSQL) DSN() string {
 	query := url.Values{}
 	query.Add("database", m.Database)
