+++ lib/config/snowflake.go
@@ -8,19 +8,6 @@ import (
 	"github.com/snowflakedb/gosnowflake"
 )
 
-type Snowflake struct {
-	AccountID string `yaml:"account"`
-	Username  string `yaml:"username"`
-	// If pathToPrivateKey is specified, the password field will be ignored
-	PathToPrivateKey string `yaml:"pathToPrivateKey,omitempty"`
-	Password         string `yaml:"password,omitempty"`
-
-	Warehouse   string `yaml:"warehouse"`
-	Region      string `yaml:"region"`
-	Host        string `yaml:"host"`
-	Application string `yaml:"application"`
-}
-
 func (s Snowflake) ToConfig() (*gosnowflake.Config, error) {
 	cfg := &gosnowflake.Config{
 		Account:     s.AccountID,
